modified updateEvents function and inputZone()
