+++ src/components/calendarComponent.js
@@ -6,36 +6,11 @@ import { useState, useEffect } from 'react';
 
 const localizer = momentLocalizer(moment);
 
-const CalendarView = () => {
-  const [events, setEvents] = useState([]);
+const CalendarView = (props) => {
+  const { events } = props;
 
-  useEffect(() => {
-    const newEvents = [];
 
-    for (let i = 0; i < localStorage.length; i++) {
-      const key = localStorage.key(i);
-      if (key.startsWith("formData_")) {
-        const evt = JSON.parse(localStorage.getItem(key));
-        evt.start = moment(evt.start).toDate();
-        evt.end = moment(evt.end).toDate();
-        newEvents.push(evt);
-      }
-    }
 
-    setEvents(newEvents);
-  }, []);
-    //loop through the items in local storage and access them to be created in the calendar section.
-
-    //Event object example
-
-  // for(let i=0; i<localStorage.length; i++){
-  //   const key = localStorage.key(i)
-  //   if(key.startsWith("formData_")){
-  //     storageObj = JSON.parse(localStorage.getItem(key))
-  //   }
-
-  //   //loop through the items in local storage and access them to be created in the calendar section.
-    
   //   //Event object example
   //   // {
   //   //   title: 'Appointment Call',
@@ -45,7 +20,7 @@ const CalendarView = () => {
   //   //events.push(eventObj)
  
   
-  
+  console.log('Events:', events);
   return (
     <div style={{ height: '600px' }}>
       <Calendar
